+++ ecdf_estimator/__init__.py
@@ -4,3 +4,6 @@ from .objective_function import multiple_objectives as multiple_estimator
 
 from .plot import *
 from .utils import *
+
+# Functions that are useful within the package, but should not be seen by the user.
+# from .auxiliaries import *
